+++ sheriff/src/Sheriff/Plugin.hs
@@ -2,8 +2,9 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE ImplicitParams #-}
-{-# LANGUAGE TypeSynonymInstances #-}
 {-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeSynonymInstances #-}
 
 module Sheriff.Plugin (plugin) where
 
@@ -175,8 +176,8 @@ sheriff opts modSummary tcEnv = do
   -- STAGE-2
   -- Get Instance declarations and add class name to module name binding in initial state
   insts <- tcg_insts . env_gbl <$> getEnv
-  let namesModTuple = fmap (\inst -> (is_dfun_name inst, getModuleName . className . is_cls $ inst)) insts
-      nameModMap = foldr (\(name, modname) r -> HM.insert (NMV_Name name) (NMV_Module modname) r) HM.empty namesModTuple
+  let namesModTuple = concatMap (\inst -> let clsName = className (is_cls inst) in (is_dfun_name inst, clsName) : fmap (\clsMethod -> (varName clsMethod, clsName)) (classMethods $ is_cls inst)) insts
+      nameModMap = foldr (\(name, clsName) r -> HM.insert (NMV_Name name) (NMV_ClassModule clsName (getModuleName clsName)) r) HM.empty namesModTuple
   
   rawErrors <- concat <$> (mapM (loopOverModBinds finalSheriffRules) $ bagToList $ tcg_binds tcEnv)
   (rawInfiniteRecursionErrors, _) <- flip runStateT nameModMap $ concat <$> (mapM (checkInfiniteRecursion True infRule) $ bagToList $ tcg_binds tcEnv)
@@ -648,7 +649,7 @@ Part-2 Validation
 -- Function to check if given DB rules is violated or not
 -- TODO: Fix this, keep two separate options for - 1. Match All Fields in AND   2. Use 1st column matching or all columns matching for composite key 
 validateDBRule :: (HasPluginOpts PluginOpts) => DBRule -> String -> [LHsExpr GhcTc] -> LHsExpr GhcTc -> TcM ([(LHsExpr GhcTc, Violation)])
-validateDBRule rule@(DBRule ruleName ruleTableName ruleColNames _ _) tableName clauses expr = do
+validateDBRule rule@(DBRule {db_rule_name = ruleName, table_name = ruleTableName, indexed_cols_names = ruleColNames}) tableName clauses expr = do
   simplifiedExprs <- trfWhereToSOP clauses
   let checkDBViolation = case (matchAllInsideAnd . pluginOpts $ ?pluginOpts) of
                           True  -> checkDBViolationMatchAll
@@ -724,14 +725,19 @@ trfWhereToSOP (clause : ls) = do
     (fn, _) -> when ((logWarnInfo . pluginOpts $ ?pluginOpts)) (liftIO $ print $ "Invalid/unknown clause in `where` clause : " <> fn <> " at " <> (showS . getLoc2 $ clause)) >> trfWhereToSOP ls
 
 -- Get table field name and table name for the `Se.Is` clause
+-- Patterns to match 'getField`, `recordDot`, `overloadedRecordDot` (ghc > 9), selector (duplicate record fields), rec fields (ghc 9), lens
 -- TODO: Refactor this to use HasField instance if possible
 getIsClauseData :: (HasPluginOpts PluginOpts) => LHsExpr GhcTc -> LHsExpr GhcTc -> LHsExpr GhcTc -> TcM (Maybe (String, String))
 getIsClauseData fieldArg _comp _clause = do
   let fieldSpecType = getDBFieldSpecType fieldArg
   mbColNameAndTableName <- case fieldSpecType of
-    None     -> when ((logWarnInfo . pluginOpts $ ?pluginOpts)) (liftIO $ print "Can't identify the way in which DB field is specified") >> pure Nothing
+    None     -> when ((logWarnInfo . pluginOpts $ ?pluginOpts)) (liftIO $ print $ "Can't identify the way in which DB field is specified: " <> showS fieldArg) >> pure Nothing
     Selector -> do
-      case (splitOn ":" $ showS fieldArg) of
+      let modFieldArg arg = case arg of
+                        (L _ (HsRecFld _ fldOcc))   -> showS $ selectorAmbiguousFieldOcc fldOcc
+                        (L loc (PatHsWrap _ wExpr)) -> modFieldArg (L loc wExpr)
+                        (L _ expr)                  -> showS expr
+      case (splitOn ":" $ modFieldArg fieldArg) of
         ("$sel" : colName : tableName : []) -> pure $ Just (colName, tableName)
         _ -> when ((logWarnInfo . pluginOpts $ ?pluginOpts)) (liftIO $ print "Invalid pattern for Selector way") >> pure Nothing
     RecordDot -> do
@@ -784,17 +790,32 @@ getIsClauseData fieldArg _comp _clause = do
                                   TyConApp ty1 _ -> showS ty1
                                   ty             -> showS ty
               pure $ Just (tail $ showS lens, take (length tblName' - 1) tblName')
+#if __GLASGOW_HASKELL__ >= 900
+            (PatHsExpansion orig (HsApp _ (L _ (HsApp _ _ tableVar)) fldVar)) -> do
+              let fldName = tail $ showS fldVar
+              typ <- getHsExprType (logTypeDebugging . pluginOpts $ ?pluginOpts) tableVar
+              let tblName' = case typ of
+                              AppTy ty1 _    -> showS ty1
+                              TyConApp ty1 _ -> showS ty1
+                              ty             -> showS ty
+              pure $ Just (fldName, take (length tblName' - 1) tblName')
+#endif                            
             _ -> when ((logWarnInfo . pluginOpts $ ?pluginOpts)) (liftIO $ putStrLn "OpApp not present. Should never be the case as we already filtered.") >> pure Nothing
   
   pure mbColNameAndTableName
 
 -- Get how DB field is being extracted in sequelize
 getDBFieldSpecType :: LHsExpr GhcTc -> DBFieldSpecType
-getDBFieldSpecType (L _ expr)
-  | isPrefixOf "$sel" (showS expr) = Selector
-  | isInfixOf "^." (showS expr) = Lens
-  | (\x -> isInfixOf "@" x) (showS expr) = RecordDot
-  | otherwise = None
+getDBFieldSpecType (L loc expr)
+  | (PatHsWrap _ wExpr) <- expr = getDBFieldSpecType (L loc wExpr)
+  | (HsRecFld _ fldOcc) <- expr = checkExprString . showS $ selectorAmbiguousFieldOcc fldOcc
+  | otherwise                   = checkExprString $ showS expr
+  where
+    checkExprString exprStr
+      | isPrefixOf "$sel" exprStr       = Selector
+      | isInfixOf "^." exprStr          = Lens
+      | (\x -> isInfixOf "@" x) exprStr = RecordDot
+      | otherwise                       = None
 
 -- Get function name for the where clause for db rules cases
 getWhereClauseFnNameWithAllArgs :: LHsExpr GhcTc -> Maybe (String, [LHsExpr GhcTc])
@@ -814,6 +835,14 @@ getWhereClauseFnNameWithAllArgs (L loc ap@(HsPar _ expr)) = getWhereClauseFnName
 -- If condition inside the list, add dummy type
 getWhereClauseFnNameWithAllArgs (L loc ap@(PatHsIf _pred thenCl elseCl)) = Just ("Or", [L loc (PatExplicitList (LitTy (StrTyLit "Dummy")) [thenCl, elseCl])])
 getWhereClauseFnNameWithAllArgs (L loc ap@(PatHsWrap _ expr)) = getWhereClauseFnNameWithAllArgs (L loc expr)
+#if __GLASGOW_HASKELL__ >= 900
+getWhereClauseFnNameWithAllArgs (L loc ap@(PatHsExpansion orig expanded)) = 
+  case (orig, expanded) of
+    ((OpApp _ _ op _), (HsApp _ (L _ (HsApp _ op' funl)) funr)) -> case showS op of
+      "($)" -> getWhereClauseFnNameWithAllArgs (L loc (HsApp noExtFieldOrAnn funl funr))
+      _ -> getWhereClauseFnNameWithAllArgs (L loc expanded)
+    _ -> getWhereClauseFnNameWithAllArgs (L loc expanded)
+#endif
 getWhereClauseFnNameWithAllArgs (L loc ap@(ExprWithTySig _ expr _)) = getWhereClauseFnNameWithAllArgs expr
 getWhereClauseFnNameWithAllArgs _ = Nothing
 
@@ -979,6 +1008,9 @@ isFunApp _ _ = False
 isLensOpApp :: HsExpr GhcTc -> Bool
 isLensOpApp (OpApp _ _ op _) = showS op == "(^.)"
 isLensOpApp (SectionR _ op _) = showS op == "(^.)"
+#if __GLASGOW_HASKELL__ >= 900
+isLensOpApp (PatHsExpansion (OpApp _ _ op _) expanded) = showS op == "(^.)"
+#endif
 isLensOpApp _ = False
 
 -- If the type is literal type, get the string name of the literal, else return the showS verison of the type